+++ test/fixtures/robot/002/inventory_rules/dl_2.py
@@ -2,20 +2,23 @@
     {
         'condition': {}, 
         'value': {
-            'robot_discovery_level': [
+            'robot_discovery_settings': [
                 (
                     'invalidpatternwillnevermatch', 
-                    '3'
+                    '3',
+                    ''
                 ), 
                 (
                     '.*', 
-                    '2'
+                    '2',
+                    ''
                 )
             ], 
             'robot_service_prefix': [
                 (
                     'invalidpatternwillnevermatch', 
-                    'FOO'
+                    'FOO',
+                    ''
                 ),
             ]
         },