+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Fixed
 
+
+## [v0.1.6 - 2021-01-05]
+
+## Added
+
+- Inventory blacklist: when using discovery level, now it is possible to blacklist certain nodes which should not be 
+  inventorized as services. (solves #54)
+
 ## [v0.1.5 - 2021-01-04]
 ### Added